project_contributors/requirements.txt
Django==3.2.9
psycopg2-binary
djangorestframework
